pyproject.toml
[project]
name = "firecrawl-skills"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "colorama>=0.4.6",
    "firecrawl-py>=2.7.1",
    "ipdb>=0.13.13",
    "ipython>=9.2.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.5",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.0",
    "typing>=3.10.0.0",
]
